29/09/25 пауль

+++ ИПМ - Пауль/17.09.25-lab1/OurDecorApplication/OurDecorApplication/Models/Material.cs
@@ -5,15 +5,14 @@ namespace OurDecorApplication.Models
     public class Material
     {
         [Key]
-        public string Id { get; set; } 
-
+        public string Id { get; set; }
         public string NameMaterial { get; set; }
-        public float? TypeMaterialId { get; set; }
+        public double? TypeMaterialId { get; set; }
         public string TypeMaterial { get; set; }
-        public float? CostMaterialOne { get; set; }
-        public float? CountMaterial { get; set; }
-        public float? MinCount { get; set; }
-        public float? CountInPackage { get; set; }
+        public double? CostMaterialOne { get; set; }
+        public double? CountMaterial { get; set; }
+        public double? MinCount { get; set; }
+        public double? CountInPackage { get; set; }
         public string UnitMeasurement { get; set; }
     }
 }
